avocado.core.tree: Avoid cyclical dependency with TreeNode.
Move TreeNode class to the new module avocado.core.tree_node,
so that we can avoid cyclical dependency on
avocado.core.tree and avocado.core.tree_debug.
